Update graphene_bands.in
Hello there is a problem with the ecutwfc value. I think, it shoulg be increased. Also I tried the lines of cade you gave for the pdos calculations, but it doesn't works for graphene and I dont really know why. I have a out of index error.

you really did a great job and this really helped me thanks a lot.
Wah.